Renamed main class from Swagger2SpringBoot to Application.
