+++ src/main/java/io/tasklist/Application.java
@@ -1,0 +1,36 @@
+package io.tasklist;
+
+import org.springframework.boot.CommandLineRunner;
+import org.springframework.boot.ExitCodeGenerator;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ComponentScan;
+
+import springfox.documentation.swagger2.annotations.EnableSwagger2;
+
+@SpringBootApplication
+@EnableSwagger2
+@ComponentScan(basePackages = "io.tasklist")
+public class Application implements CommandLineRunner {
+
+	@Override
+	public void run(String... arg0) throws Exception {
+		if (arg0.length > 0 && arg0[0].equals("exitcode")) {
+			throw new ExitException();
+		}
+	}
+
+	public static void main(String[] args) throws Exception {
+		new SpringApplication(Application.class).run(args);
+	}
+
+	class ExitException extends RuntimeException implements ExitCodeGenerator {
+		private static final long serialVersionUID = 1L;
+
+		@Override
+		public int getExitCode() {
+			return 10;
+		}
+
+	}
+}
